secure_version/sharepy/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pyjwt==2.8.0
minio==7.2.0
python-dotenv==1.0.0
bcrypt==4.1.1